+++ packages/dart_open_fetch_runtime/lib/dart_open_fetch_runtime.dart
@@ -3,7 +3,9 @@ library;
 
 export 'src/api_exception.dart';
 export 'src/api_response.dart';
+export 'src/bearer_auth_middleware.dart';
 export 'src/http_adapter.dart';
+export 'src/http_client_adapter.dart';
 export 'src/http_request.dart';
 export 'src/http_response.dart';
 export 'src/middleware.dart';